requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
flasgger==0.9.7.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
authlib==1.2.1
requests==2.31.0
